updated hubspot schema

+++ models/staging/hubspot/campaigns_dbt.sql
@@ -25,7 +25,7 @@ SELECT
     _airbyte_data ->> 'name' as name,
     _airbyte_data ->> 'appName' as appname,
     _airbyte_data ->> 'subject' as subject,
-    _airbyte_data ->> 'lastUpdatedTime' as lastupdatedtime
+    (_airbyte_data ->> 'lastUpdatedTime')::bigint as lastupdatedtime
 FROM {{source ('hubspot', '_airbyte_raw_campaigns')}}
 )
